==> 00-Pipeline.ipynb <==
# Use this notebook to runn all the notebooks as a pipeline

Building this dataset takes multiple steps:

1. First we need to gather the list of pages we can use for building it, there's a search engine that returns a `json` file with the list of pages.
2. We import all the HTML and do an initial parsing to extract the debates and name of the politician for each speech.
3. Then we build a huge dataset with all the imported data, and enhance it by extracting the polititian title if any, extract sentiment and cleanup data.
4. Once we have the full dataset, we prepare it for publishing it on Hugging Face and make it compatible with the Transformers platform.

Each of this steps corresponds to a Notebook, you can run them all with this pipeline

## Step 1: run search.

The result is a bunch of json files pointing to the real HTML files, into the ./data/terms folder. We also store partial csv and a full one.

%run ./01-CDLD-Get-full-term-in-office.ipynb

## Step 2: Import all the debates from HTML

The result is stored as a csv file for each debate in ./data/debates. Those files will contain all the speeches with the name of the politician giving the speech.
We also store the HTML pages into ./data/pagecache to avoid downloading all the content each time.

%run ./02-CDLD-import.ipynb

## Step 3: Prepare and clean the dataset.

Data enhancement by removing paranthesized text and extracting data from these items like politician title, related decrees or reactions to have a sentiment score.

%run ./03-Prepare-Data.ipynb

## Step 4: NLP dataset preparation

%run ./04-Prepare-dataset.ipynb
