==> 03-Prepare-Data.ipynb <==
# Step 03: Data preparation

In this step we enrich the dataset with some information we can find in the text, like:

* When a person has a special position in the government or in the congress (President of the congress, President of the nation, Minister,...), they usually have the position name written in ALL CAPS and then their name follows it between parens. Like this: `PRESIDENTE DEL GOBIERNO (Mariano Rajoy)`
* There's also a process to enrich the dataset with additional information based on paranthesized text.


import pandas as pd
import re

files = !ls data/debates/*.csv

full = pd.concat([pd.read_csv(f) for f in files],ignore_index=True)

# constants
title_pat=r'(.*)\((.*)\)'

# find patterns like PRESIDENTE DEL GOBIERNO (Mariano Rajoy)

# extract title, ex: "PRESIDENTE DEL GOBIERNO"
full["Title"]=full["Name"].apply(lambda x: re.match(title_pat, x).groups()[0] if re.match(title_pat, x)!=None else None)
# extract name, ex: "Mariano Rajoy"
full["Name"]=full["Name"].apply(lambda x: re.match(title_pat, x).groups()[1].upper() if re.match(title_pat, x)!=None else x)
full[full.Title.notnull()].head()

import datetime
# convert to proper date
full["Date"]=full["Date"].apply(lambda f: str(f))
full.Date=full.Date.apply(lambda d: datetime.date(int(d[0:4]),int(d[4:6]),int(d[6:8])))


# find special one "CANDIDATO" because it is shifted and we have to unshift it
full["temp"]=full[full.Name.str.contains("CANDIDATO")].Name
full.loc[full["temp"].str.contains("CANDIDATO",na=False),"Name"]=full[full.Name.str.contains("CANDIDATO")].Title
full.loc[full["temp"].str.contains("CANDIDATO",na=False),"Title"]=full["temp"]
del full["temp"]
full.head(20)

# find paranthesized values and sort them by occurence
parenthesized_pat=r'(\(.+?\))'
# can also use negative matching instead of non-greedy parenthesized_pat=r'.*(\([^\)]+\))'
l=full["Text"].apply(lambda t: ",".join(re.findall(parenthesized_pat,str(t))))
pd.DataFrame({'test':l}).groupby(['test']).size().reset_index().sort_values(0,ascending=False).head(20)

# This is a test to check the paranthesized test it founds
parenthesized_pat=r'\(.+?\)'
unclosed_parenthesized_pat=r'(\(.+?\.)'
m=full["Text"].str.replace(parenthesized_pat,'',regex=True).str.strip()
l=m.apply(lambda t: ",".join([par+")" for par in re.findall(unclosed_parenthesized_pat,str(t))]))
#pd.DataFrame({'test':l}).groupby(['test']).size().reset_index().sort_values(0,ascending=False).head(20)
tdf=pd.DataFrame({'m':l})
tdf["m"].apply(lambda m: m.replace(".)",")") ).iloc[6065]



# find paranthesized values and sort them by occurence
parenthesized_pat=r'(\(.+?\))'
unclosed_parenthesized_pat=r'(\(.+?\.)'

t=[]
# can also use negative matching instead of non-greedy parenthesized_pat=r'.*(\([^\)]+\))'
for item in full["Text"]:
    items=re.findall(parenthesized_pat,str(item))
    if(len(items)>0):
        t.append(items)

for item in full["Text"].str.replace(parenthesized_pat,'',regex=True).str.strip():
    items=re.findall(unclosed_parenthesized_pat,str(item))
    if(len(items)>0):
        t.append(items)

flattened_list = [value for sub_list in t for value in sub_list]


parens=pd.DataFrame(flattened_list, columns=["Text"])
parens["Text"]=parens["Text"].str.replace(r"\.$",")")
parens["Text"]=parens["Text"].str.replace(r"\.\)",")")
parens["Text"]=parens["Text"].str.lower()
p=parens.groupby("Text").size().reset_index().sort_values(0,ascending=False)
p[:40]

# There are two types of parantesized text, the closed and the unclosed.
# We will extract  the closed ones first
parenthesized_pat=r'\(.+?\)'
# And after removing the closed we will seek for unclosed ones
unclosed_parenthesized_pat=r'(\(.+?\.)'

# Add the paranthesized text to the infos
full["infos"]=full["Text"].apply(lambda t: ",".join(re.findall(parenthesized_pat,str(t))).strip())
# now remove parenthesized sentences
full["Text"]=full["Text"].str.replace(parenthesized_pat,'',regex=True).str.strip()
# Add the unclosed sentences
full["infos2"]=full["Text"].apply( lambda t: ",".join([par+")" for par in re.findall(unclosed_parenthesized_pat,str(t))]))
# remove them
full["Text"]=full["Text"].str.replace(unclosed_parenthesized_pat,'',regex=True).str.strip()
# join the two results with comma separated ones into the infos column
full["infos"]=full[["infos","infos2"]].apply( lambda row: row["infos"] if len(row["infos2"])==0 else row["infos2"] if len(row["infos"])==0 else ",".join([row["infos"],row["infos2"]]),axis=1)
full["infos"]=full["infos"].str.replace(".)",")",regex=False)
del full["infos2"]
full["TextLen"]=full.Text.str.len()
full.describe()

# extract records
full["Data"]=full["infos"].apply( lambda f: re.search(r"mero de expediente (\d+/\d+)",f).groups()[0] if re.search(r"mero de expediente (\d+/\d+)",f) else None)
full["InterventionType"]=full["infos"].apply( lambda f: "Record" if re.search(r"mero de expediente (\d+/\d+)",f) else None)


# extract congress sentiment from interjections

full["Positive"]=0
full["Negative"]=0
full["Surprise"]=0

positives=[r"aplauso", r"risa", r"muy bien",r"si+!",r"sí",r"asentimiento"]
negatives=[r"rumore",r"no+!",r"protesta",r"denegaciones",r"qué barbaridad",r"mentira"]
surprise=[r"hala",r"oh!",r"ah!"]

for pos in positives:
    full["Positive"]+= full["infos"].apply( lambda f: len(re.findall(pos,f.lower())) if re.search(pos,f.lower()) else 0)

for pos in negatives:
    full["Negative"]+= full["infos"].apply( lambda f: len(re.findall(pos,f.lower())) if re.search(pos,f.lower()) else 0)

for pos in surprise:
    full["Surprise"]+= full["infos"].apply( lambda f: len(re.findall(pos,f.lower())) if re.search(pos,f.lower()) else 0)


full.sort_values("Positive", ascending=False)


print(full.describe())

# Store the full debates to a parquet file
full.to_parquet('../debates.parquet', index=False)
print("debates written to parquet file.")

# Find specific politicians and store them as CSV
full[full['Name'].str.contains('EZ ZAPATERO')].to_csv('../zapatero.csv', index=False)
full[full['Name'].str.contains('RAJOY')].to_csv('../rajoy.csv',index=False)

print("Exported Zapater & Rajoy")

full[full['Name'].str.contains('RAJOY')].iloc[1]["Text"]

print(full[full["Name"].str.contains("RAJOY")].describe())
