==> 05-Transformer.ipynb <==
# Fine tuning the model

* https://towardsdatascience.com/controlling-text-generation-from-language-models-6334935e80cf
 https://huggingface.co/DeepESP/gpt2-spanish
* https://huggingface.co/docs/transformers/training
* https://medium.com/nlplanet/a-full-guide-to-finetuning-t5-for-text2text-and-building-a-demo-with-streamlit-c72009631887
* finetuning a prompt model https://www.youtube.com/watch?v=uE0_XKh2d6g


from transformers import AutoTokenizer

tokenizer = AutoTokenizer.from_pretrained("bert-base-cased")

from datasets import load_dataset

ds=load_dataset('.',data_files={'train': ['../rajoy_train.jsonl'],
                             'test':['../rajoy_test.jsonl']})

def tokenize_function(examples):
    return tokenizer(examples["Text"], padding="max_length", truncation=True)

ds=ds.map(tokenize_function,batched=True)

ds

from transformers import AutoModelForSequenceClassification

model = AutoModelForSequenceClassification.from_pretrained("bert-base-cased", num_labels=5)

from transformers import TrainingArguments, Trainer

training_args = TrainingArguments(output_dir="test_trainer", evaluation_strategy="epoch")

import numpy as np
import evaluate

metric = evaluate.load("accuracy")

def compute_metrics(eval_pred):
    logits, labels = eval_pred
    predictions = np.argmax(logits, axis=-1)
    return metric.compute(predictions=predictions, references=labels)

ds['train']

trainer = Trainer(
    model=model,
    args=training_args,
    train_dataset=ds['train'],
    eval_dataset=ds['test'],
#    compute_metrics=compute_metrics,
)

trainer.train()
